requirements.txt
tensorflow==1.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Pillow==7.1.2
numpy==1.18.2
streamlit==0.71.0
protobuf==3.14.0
opencv-contrib-python==4.2.0.34
opencv-python==4.4.0.40
